aws_cdk_cloudwatch/requirements.txt
-e .
aws.cdk.aws_ec2
aws.cdk.aws_iam
aws.cdk.aws_lambda
aws.cdk.aws_sns
aws.cdk.aws_sns_subscriptions
aws.cdk.aws_elasticloadbalancingv2
aws.cdk.aws_autoscaling
aws.cdk.aws_cloudwatch
aws.cdk.aws_cloudwatch_actions
